revm_database::in_memory_db

Struct CacheDB

pub struct CacheDB<ExtDB> {
    pub cache: Cache,
    pub db: ExtDB,
}

A Database implementation that stores all state changes in memory.

This implementation wraps a DatabaseRef that is used to load data (AccountInfo).

Fields
cache: Cache

The cache that stores all state changes.

db: ExtDB

The underlying database (DatabaseRef) that is used to load data.

Note: This is read-only, data is never written to this database.

Implementations

impl<ExtDb> CacheDB<CacheDB<ExtDb>>

pub fn flatten(self) -> CacheDB<ExtDb>

Flattens a nested cache by applying the outer cache to the inner cache.

The behavior is as follows:

Accounts are overridden with outer accounts

Contracts are overridden with outer contracts

Logs are appended

Block hashes are overridden with outer block hashes

pub fn discard_outer(self) -> CacheDB<ExtDb>

Discards the outer cache and return the inner cache.

impl<ExtDB> CacheDB<ExtDB>

pub fn new(db: ExtDB) -> Self

Creates a new cache with the given external database.

pub fn insert_contract(&mut self, account: &mut AccountInfo)

Inserts the account’s code into the cache.

Accounts objects and code are stored separately in the cache, this will take the code from the account and instead map it to the code hash.

Note: This will not insert into the underlying external database.

pub fn insert_account_info(&mut self, address: Address, info: AccountInfo)

Inserts account info but not override storage

pub fn nest(self) -> CacheDB<Self>

Wraps the cache in a CacheDB, creating a nested cache.

impl<ExtDB: DatabaseRef> CacheDB<ExtDB>

pub fn load_account( &mut self, address: Address, ) -> Result<&mut DbAccount, ExtDB::Error>

Returns the account for the given address.

If the account was not found in the cache, it will be loaded from the underlying database.

pub fn insert_account_storage( &mut self, address: Address, slot: StorageKey, value: StorageValue, ) -> Result<(), ExtDB::Error>

Inserts account storage without overriding account info

pub fn replace_account_storage( &mut self, address: Address, storage: HashMap<StorageKey, StorageValue>, ) -> Result<(), ExtDB::Error>

Replaces account storage without overriding account info