revm_database::states::state

Struct State

pub struct State<DB> {
    pub cache: CacheState,
    pub database: DB,
    pub transition_state: Option<TransitionState>,
    pub bundle_state: BundleState,
    pub use_preloaded_bundle: bool,
    pub block_hashes: BTreeMap<u64, B256>,
}

State of blockchain

State clear flag is set inside CacheState and by default it is enabled.

If you want to disable it use set_state_clear_flag function.

Fields
cache: CacheState

Cached state contains both changed from evm execution and cached/loaded account/storages from database

This allows us to have only one layer of cache where we can fetch data.

Additionally, we can introduce some preloading of data from database.

database: DB

Optional database that we use to fetch data from

If database is not present, we will return not existing account and storage.

Note: It is marked as Send so database can be shared between threads.

transition_state: Option<TransitionState>

Block state, it aggregates transactions transitions into one state

Build reverts and state that gets applied to the state.

bundle_state: BundleState

After block is finishes we merge those changes inside bundle

Bundle is used to update database and create changesets.

Bundle state can be set on initialization if we want to use preloaded bundle.

use_preloaded_bundle: bool

Addition layer that is going to be used to fetched values before fetching values from database

Bundle is the main output of the state execution and this allows setting previous bundle and using its values for execution.

block_hashes: BTreeMap<u64, B256>

If EVM asks for block hash, we will first check if they are found here, then ask the database

This map can be used to give different values for block hashes if in case.

The fork block is different or some blocks are not saved inside database.

Implementations

impl State<EmptyDB>

pub fn builder() -> StateBuilder<EmptyDB>

Return the builder that build the State.

impl<DB: Database> State<DB>

pub fn bundle_size_hint(&self) -> usize

Returns the size hint for the inner bundle state.

See BundleState::size_hint for more info.

pub fn increment_balances( &mut self, balances: impl IntoIterator<Item = (Address, u128)>, ) -> Result<(), DB::Error>

Iterates over received balances and increment all account balances.

Note: If account is not found inside cache state it will be loaded from database.

Update will create transitions for all accounts that are updated.

If using this to implement withdrawals, zero balances must be filtered out before calling this function.

pub fn drain_balances( &mut self, addresses: impl IntoIterator<Item = Address>, ) -> Result<Vec<u128>, DB::Error>

Drains balances from given account and return those values.

It is used for DAO hardfork state change to move values from given accounts.

pub fn set_state_clear_flag(&mut self, has_state_clear: bool)

State clear EIP-161 is enabled in Spurious Dragon hardfork.

pub fn insert_not_existing(&mut self, address: Address)

pub fn insert_account(&mut self, address: Address, info: AccountInfo)

pub fn insert_account_with_storage( &mut self, address: Address, info: AccountInfo, storage: PlainStorage, )

pub fn apply_transition( &mut self, transitions: Vec<(Address, TransitionAccount)>, )

Applies evm transitions to transition state.

pub fn merge_transitions(&mut self, retention: BundleRetention)

Take all transitions and merge them inside bundle state.

This action will create final post state and all reverts so that we at any time revert state of bundle to the state before transition is applied.

pub fn load_cache_account( &mut self, address: Address, ) -> Result<&mut CacheAccount, DB::Error>

Get a mutable reference to the CacheAccount for the given address.

If the account is not found in the cache, it will be loaded from the database and inserted into the cache.

pub fn take_bundle(&mut self) -> BundleState

Takess the BundleState changeset from the State, replacing it with an empty one.

This will not apply any pending TransitionState.

It is recommended to call State::merge_transitions before taking the bundle.

If the State has been built with the StateBuilder::with_bundle_prestate option, the pre-state will be taken along with any changes made by State::merge_transitions.