revm_database

Trait TryDatabaseCommit

pub trait TryDatabaseCommit {
    type Error: Error;

    // Required method
    fn try_commit(
        &mut self,
        changes: HashMap<Address, Account>,
    ) -> Result<(), Self::Error>;
}

EVM database commit interface that can fail.

This is intended for use with types that may fail to commit changes, e.g. because they are directly interacting with the filesystem, or must arrange access to a shared resource.

Required Associated Types

type Error: Error

Error type for when TryDatabaseCommit::try_commit fails.

Required Methods

fn try_commit( &mut self, changes: HashMap<Address, Account>, ) -> Result<(), Self::Error>

Attempt to commit changes to the database.