Crate revm_database

Database implementations.

Re-exports

pub use states::AccountRevert;

pub use states::AccountStatus;

pub use states::BundleAccount;

pub use states::BundleState;

pub use states::CacheState;

pub use states::DBBox;

pub use states::OriginalValuesKnown;

pub use states::PlainAccount;

pub use states::RevertToSlot;

pub use states::State;

pub use states::StateBuilder;

pub use states::StateDBBox;

pub use states::StorageWithOriginalValues;

pub use states::TransitionAccount;

pub use states::TransitionState;

pub use in_memory_db::*;

Modules

empty_db

in_memory_db

states

try_commit

Structs

ArcUpgradeError

Error type for implementation of TryDatabaseCommit on Arc.

EmptyDBTyped

An empty database that always returns default values when queried

WrapDatabaseRef

Wraps a DatabaseRef to provide a Database implementation.

Traits

DBErrorMarker

Database error marker is needed to implement From conversion for Error type.

Database

EVM database interface.

DatabaseCommit

EVM database commit interface.

DatabaseRef

EVM database interface.

TryDatabaseCommit

EVM database commit interface that can fail.

Type Aliases

EmptyDB

An empty database that always returns default values when queried